Added DefaultApiConventions

+++ StockHubApi/StockHubApi/Controllers/StockController.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using Microsoft.AspNetCore.Mvc;
 using StockHubApi.Models;
 using StockHubApi.Services;
@@ -28,9 +30,17 @@ namespace StockHubApi.Controllers
         /// </summary>
         /// <returns>An <see cref="IEnumerable{T}" /> of <see cref="Stock" />.</returns>
         [HttpGet]
-        public IEnumerable<Stock> Index()
+        [ApiConventionMethod(typeof(DefaultApiConventions), nameof(DefaultApiConventions.Get))]
+        public IActionResult Index()
         {
-            return stockService.GetStocks();
+            IEnumerable<Stock> stocks = stockService.GetStocks();
+
+            if (stocks == null || !stocks.Any())
+            {
+                return NotFound();
+            }
+
+            return Ok(stocks);
         }
 
         /// <summary>
@@ -39,9 +49,17 @@ namespace StockHubApi.Controllers
         /// <param name="id">The id of the a <see cref="Stock" /> which will be retreived.</param>
         /// <returns>A <see cref="Stock" /> by the given id.</returns>
         [HttpGet("{id}")]
-        public Stock Show(int id)
+        [ApiConventionMethod(typeof(DefaultApiConventions), nameof(DefaultApiConventions.Get))]
+        public IActionResult Show(int id)
         {
-            return stockService.GetStock(id);
+            Stock stock = stockService.GetStock(id);
+
+            if (stock == null)
+            {
+                return NotFound();
+            }
+
+            return Ok(stock);
         }
 
         /// <summary>
@@ -49,9 +67,11 @@ namespace StockHubApi.Controllers
         /// </summary>
         /// <param name="stock">A instance of a <see cref="Stock" /> which contains which will be created.</param>
         [HttpPost]
-        public void Create(Stock stock)
+        [ApiConventionMethod(typeof(DefaultApiConventions), nameof(DefaultApiConventions.Create))]
+        public IActionResult Create(Stock stock)
         {
-            stockService.CreateStock(stock);
+            Stock dbStock = stockService.CreateStock(stock);
+            return Created(new Uri("/Stock", UriKind.Relative), dbStock);
         }
 
         /// <summary>
@@ -59,9 +79,11 @@ namespace StockHubApi.Controllers
         /// </summary>
         /// <param name="stock">A instance of a <see cref="Stock" /> which will be updated with the given values.</param>
         [HttpPut("{id}")]
-        public void Update(Stock stock)
+        [ApiConventionMethod(typeof(DefaultApiConventions), nameof(DefaultApiConventions.Put))]
+        public IActionResult Update(Stock stock)
         {
             stockService.UpdateStock(stock);
+            return NoContent();
         }
 
         /// <summary>
@@ -69,9 +91,19 @@ namespace StockHubApi.Controllers
         /// </summary>
         /// <param name="id">The id of the a <see cref="Stock" /> which will be deleted.</param>
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        [ApiConventionMethod(typeof(DefaultApiConventions), nameof(DefaultApiConventions.Delete))]
+        public IActionResult Delete(int id)
         {
+            Stock databaseStock = stockService.GetStock(id);
+
+            if (databaseStock == null)
+            {
+                return NotFound();
+            }
+
             stockService.DeleteStock(id);
+
+            return Ok();
         }
     }
 }
